Move protos package to internal
The compiled protos are not intended for use by external consumers, and
are shared across both the client and the server.

[#158534897]
