+++ internal/protos/protosfakes/fake_permission_service_server.go
@@ -4,19 +4,17 @@ package protosfakes
 import (
 	"sync"
 
-	"code.cloudfoundry.org/perm/api/protos"
+	"code.cloudfoundry.org/perm/internal/protos"
 	_ "github.com/gogo/protobuf/gogoproto"
 	"golang.org/x/net/context"
-	"google.golang.org/grpc"
 )
 
-type FakePermissionServiceClient struct {
-	HasPermissionStub        func(ctx context.Context, in *protos.HasPermissionRequest, opts ...grpc.CallOption) (*protos.HasPermissionResponse, error)
+type FakePermissionServiceServer struct {
+	HasPermissionStub        func(context.Context, *protos.HasPermissionRequest) (*protos.HasPermissionResponse, error)
 	hasPermissionMutex       sync.RWMutex
 	hasPermissionArgsForCall []struct {
-		ctx  context.Context
-		in   *protos.HasPermissionRequest
-		opts []grpc.CallOption
+		arg1 context.Context
+		arg2 *protos.HasPermissionRequest
 	}
 	hasPermissionReturns struct {
 		result1 *protos.HasPermissionResponse
@@ -26,12 +24,11 @@ type FakePermissionServiceClient struct {
 		result1 *protos.HasPermissionResponse
 		result2 error
 	}
-	ListResourcePatternsStub        func(ctx context.Context, in *protos.ListResourcePatternsRequest, opts ...grpc.CallOption) (*protos.ListResourcePatternsResponse, error)
+	ListResourcePatternsStub        func(context.Context, *protos.ListResourcePatternsRequest) (*protos.ListResourcePatternsResponse, error)
 	listResourcePatternsMutex       sync.RWMutex
 	listResourcePatternsArgsForCall []struct {
-		ctx  context.Context
-		in   *protos.ListResourcePatternsRequest
-		opts []grpc.CallOption
+		arg1 context.Context
+		arg2 *protos.ListResourcePatternsRequest
 	}
 	listResourcePatternsReturns struct {
 		result1 *protos.ListResourcePatternsResponse
@@ -45,18 +42,17 @@ type FakePermissionServiceClient struct {
 	invocationsMutex sync.RWMutex
 }
 
-func (fake *FakePermissionServiceClient) HasPermission(ctx context.Context, in *protos.HasPermissionRequest, opts ...grpc.CallOption) (*protos.HasPermissionResponse, error) {
+func (fake *FakePermissionServiceServer) HasPermission(arg1 context.Context, arg2 *protos.HasPermissionRequest) (*protos.HasPermissionResponse, error) {
 	fake.hasPermissionMutex.Lock()
 	ret, specificReturn := fake.hasPermissionReturnsOnCall[len(fake.hasPermissionArgsForCall)]
 	fake.hasPermissionArgsForCall = append(fake.hasPermissionArgsForCall, struct {
-		ctx  context.Context
-		in   *protos.HasPermissionRequest
-		opts []grpc.CallOption
-	}{ctx, in, opts})
-	fake.recordInvocation("HasPermission", []interface{}{ctx, in, opts})
+		arg1 context.Context
+		arg2 *protos.HasPermissionRequest
+	}{arg1, arg2})
+	fake.recordInvocation("HasPermission", []interface{}{arg1, arg2})
 	fake.hasPermissionMutex.Unlock()
 	if fake.HasPermissionStub != nil {
-		return fake.HasPermissionStub(ctx, in, opts...)
+		return fake.HasPermissionStub(arg1, arg2)
 	}
 	if specificReturn {
 		return ret.result1, ret.result2
@@ -64,19 +60,19 @@ func (fake *FakePermissionServiceClient) HasPermission(ctx context.Context, in *
 	return fake.hasPermissionReturns.result1, fake.hasPermissionReturns.result2
 }
 
-func (fake *FakePermissionServiceClient) HasPermissionCallCount() int {
+func (fake *FakePermissionServiceServer) HasPermissionCallCount() int {
 	fake.hasPermissionMutex.RLock()
 	defer fake.hasPermissionMutex.RUnlock()
 	return len(fake.hasPermissionArgsForCall)
 }
 
-func (fake *FakePermissionServiceClient) HasPermissionArgsForCall(i int) (context.Context, *protos.HasPermissionRequest, []grpc.CallOption) {
+func (fake *FakePermissionServiceServer) HasPermissionArgsForCall(i int) (context.Context, *protos.HasPermissionRequest) {
 	fake.hasPermissionMutex.RLock()
 	defer fake.hasPermissionMutex.RUnlock()
-	return fake.hasPermissionArgsForCall[i].ctx, fake.hasPermissionArgsForCall[i].in, fake.hasPermissionArgsForCall[i].opts
+	return fake.hasPermissionArgsForCall[i].arg1, fake.hasPermissionArgsForCall[i].arg2
 }
 
-func (fake *FakePermissionServiceClient) HasPermissionReturns(result1 *protos.HasPermissionResponse, result2 error) {
+func (fake *FakePermissionServiceServer) HasPermissionReturns(result1 *protos.HasPermissionResponse, result2 error) {
 	fake.HasPermissionStub = nil
 	fake.hasPermissionReturns = struct {
 		result1 *protos.HasPermissionResponse
@@ -84,7 +80,7 @@ func (fake *FakePermissionServiceClient) HasPermissionReturns(result1 *protos.Ha
 	}{result1, result2}
 }
 
-func (fake *FakePermissionServiceClient) HasPermissionReturnsOnCall(i int, result1 *protos.HasPermissionResponse, result2 error) {
+func (fake *FakePermissionServiceServer) HasPermissionReturnsOnCall(i int, result1 *protos.HasPermissionResponse, result2 error) {
 	fake.HasPermissionStub = nil
 	if fake.hasPermissionReturnsOnCall == nil {
 		fake.hasPermissionReturnsOnCall = make(map[int]struct {
@@ -98,18 +94,17 @@ func (fake *FakePermissionServiceClient) HasPermissionReturnsOnCall(i int, resul
 	}{result1, result2}
 }
 
-func (fake *FakePermissionServiceClient) ListResourcePatterns(ctx context.Context, in *protos.ListResourcePatternsRequest, opts ...grpc.CallOption) (*protos.ListResourcePatternsResponse, error) {
+func (fake *FakePermissionServiceServer) ListResourcePatterns(arg1 context.Context, arg2 *protos.ListResourcePatternsRequest) (*protos.ListResourcePatternsResponse, error) {
 	fake.listResourcePatternsMutex.Lock()
 	ret, specificReturn := fake.listResourcePatternsReturnsOnCall[len(fake.listResourcePatternsArgsForCall)]
 	fake.listResourcePatternsArgsForCall = append(fake.listResourcePatternsArgsForCall, struct {
-		ctx  context.Context
-		in   *protos.ListResourcePatternsRequest
-		opts []grpc.CallOption
-	}{ctx, in, opts})
-	fake.recordInvocation("ListResourcePatterns", []interface{}{ctx, in, opts})
+		arg1 context.Context
+		arg2 *protos.ListResourcePatternsRequest
+	}{arg1, arg2})
+	fake.recordInvocation("ListResourcePatterns", []interface{}{arg1, arg2})
 	fake.listResourcePatternsMutex.Unlock()
 	if fake.ListResourcePatternsStub != nil {
-		return fake.ListResourcePatternsStub(ctx, in, opts...)
+		return fake.ListResourcePatternsStub(arg1, arg2)
 	}
 	if specificReturn {
 		return ret.result1, ret.result2
@@ -117,19 +112,19 @@ func (fake *FakePermissionServiceClient) ListResourcePatterns(ctx context.Contex
 	return fake.listResourcePatternsReturns.result1, fake.listResourcePatternsReturns.result2
 }
 
-func (fake *FakePermissionServiceClient) ListResourcePatternsCallCount() int {
+func (fake *FakePermissionServiceServer) ListResourcePatternsCallCount() int {
 	fake.listResourcePatternsMutex.RLock()
 	defer fake.listResourcePatternsMutex.RUnlock()
 	return len(fake.listResourcePatternsArgsForCall)
 }
 
-func (fake *FakePermissionServiceClient) ListResourcePatternsArgsForCall(i int) (context.Context, *protos.ListResourcePatternsRequest, []grpc.CallOption) {
+func (fake *FakePermissionServiceServer) ListResourcePatternsArgsForCall(i int) (context.Context, *protos.ListResourcePatternsRequest) {
 	fake.listResourcePatternsMutex.RLock()
 	defer fake.listResourcePatternsMutex.RUnlock()
-	return fake.listResourcePatternsArgsForCall[i].ctx, fake.listResourcePatternsArgsForCall[i].in, fake.listResourcePatternsArgsForCall[i].opts
+	return fake.listResourcePatternsArgsForCall[i].arg1, fake.listResourcePatternsArgsForCall[i].arg2
 }
 
-func (fake *FakePermissionServiceClient) ListResourcePatternsReturns(result1 *protos.ListResourcePatternsResponse, result2 error) {
+func (fake *FakePermissionServiceServer) ListResourcePatternsReturns(result1 *protos.ListResourcePatternsResponse, result2 error) {
 	fake.ListResourcePatternsStub = nil
 	fake.listResourcePatternsReturns = struct {
 		result1 *protos.ListResourcePatternsResponse
@@ -137,7 +132,7 @@ func (fake *FakePermissionServiceClient) ListResourcePatternsReturns(result1 *pr
 	}{result1, result2}
 }
 
-func (fake *FakePermissionServiceClient) ListResourcePatternsReturnsOnCall(i int, result1 *protos.ListResourcePatternsResponse, result2 error) {
+func (fake *FakePermissionServiceServer) ListResourcePatternsReturnsOnCall(i int, result1 *protos.ListResourcePatternsResponse, result2 error) {
 	fake.ListResourcePatternsStub = nil
 	if fake.listResourcePatternsReturnsOnCall == nil {
 		fake.listResourcePatternsReturnsOnCall = make(map[int]struct {
@@ -151,7 +146,7 @@ func (fake *FakePermissionServiceClient) ListResourcePatternsReturnsOnCall(i int
 	}{result1, result2}
 }
 
-func (fake *FakePermissionServiceClient) Invocations() map[string][][]interface{} {
+func (fake *FakePermissionServiceServer) Invocations() map[string][][]interface{} {
 	fake.invocationsMutex.RLock()
 	defer fake.invocationsMutex.RUnlock()
 	fake.hasPermissionMutex.RLock()
@@ -165,7 +160,7 @@ func (fake *FakePermissionServiceClient) Invocations() map[string][][]interface{
 	return copiedInvocations
 }
 
-func (fake *FakePermissionServiceClient) recordInvocation(key string, args []interface{}) {
+func (fake *FakePermissionServiceServer) recordInvocation(key string, args []interface{}) {
 	fake.invocationsMutex.Lock()
 	defer fake.invocationsMutex.Unlock()
 	if fake.invocations == nil {
@@ -177,4 +172,4 @@ func (fake *FakePermissionServiceClient) recordInvocation(key string, args []int
 	fake.invocations[key] = append(fake.invocations[key], args)
 }
 
-var _ protos.PermissionServiceClient = new(FakePermissionServiceClient)
+var _ protos.PermissionServiceServer = new(FakePermissionServiceServer)
